+++ application/controllers/Jitsi_settings.php
@@ -1,0 +1,95 @@
+<?php defined('BASEPATH') or exit('No direct script access allowed');
+
+/* ----------------------------------------------------------------------------
+ * Easy!Appointments - Online Appointment Scheduler
+ *
+ * @package     EasyAppointments
+ * @author      A.Tselegidis <alextselegidis@gmail.com>
+ * @copyright   Copyright (c) Alex Tselegidis
+ * @license     https://opensource.org/licenses/GPL-3.0 - GPLv3
+ * @link        https://easyappointments.org
+ * @since       v1.6.0
+ * ---------------------------------------------------------------------------- */
+
+/**
+ * Jitsi_settings controller.
+ *
+ * Handles Jitsi integration settings.
+ *
+ * @package Controllers
+ */
+class Jitsi_settings extends EA_Controller
+{
+    /**
+     * Jitsi_settings constructor.
+     */
+    public function __construct()
+    {
+        parent::__construct();
+
+        $this->load->model('roles_model');
+
+        $role_slug = session('role_slug');
+
+        $required_permissions = can('edit', PRIV_SYSTEM_SETTINGS);
+
+        if ($required_permissions === false) {
+            show_error('Forbidden', 403);
+        }
+    }
+
+    /**
+     * Render the settings page.
+     */
+    public function index(): void
+    {
+        $user_id = session('user_id');
+
+        $role_slug = session('role_slug');
+
+        $jitsi_settings = [
+            [
+                'name' => 'jitsi_enabled',
+                'value' => setting('jitsi_enabled', '0'),
+            ],
+        ];
+
+        script_vars([
+            'user_id' => $user_id,
+            'role_slug' => $role_slug,
+            'jitsi_settings' => $jitsi_settings,
+        ]);
+
+        html_vars([
+            'page_title' => lang('settings'),
+            'active_menu' => PRIV_SYSTEM_SETTINGS,
+            'user_display_name' => $this->accounts->get_user_display_name($user_id),
+        ]);
+
+        $this->load->view('pages/jitsi_settings');
+    }
+
+    /**
+     * Save the Jitsi settings.
+     */
+    public function save(): void
+    {
+        try {
+            if (cannot('edit', PRIV_SYSTEM_SETTINGS)) {
+                abort(403, 'Forbidden');
+            }
+
+            $jitsi_settings = request('jitsi_settings', []);
+
+            foreach ($jitsi_settings as $jitsi_setting) {
+                setting([
+                    $jitsi_setting['name'] => $jitsi_setting['value'],
+                ]);
+            }
+
+            response();
+        } catch (Throwable $e) {
+            json_exception($e);
+        }
+    }
+}