+++ CHANGELOG.md
@@ -8,6 +8,7 @@ developers to maintain and readjust their custom modifications on the main proje
 ### Added
 
 - Added request method check on each request so that only allowed methods are accepted (-)
+- Add Jitsi integration and link generation for appointments made via the public page (-)
 
 ### Changed